chore: refactor test utils
