+++ drivers/oracle/internal/performance_test.go
@@ -4,14 +4,14 @@ import (
 	"context"
 	"testing"
 
-	"github.com/datazip-inc/olake/drivers/abstract"
 	"github.com/datazip-inc/olake/utils"
+	"github.com/datazip-inc/olake/utils/testutils"
 	"github.com/jmoiron/sqlx"
 )
 
 func TestOraclePerformance(t *testing.T) {
-	config := abstract.PerformanceTestConfig{
-		TestConfig:      abstract.GetTestConfig("oracle"),
+	config := testutils.PerformanceTestConfig{
+		TestConfig:      testutils.GetTestConfig("oracle"),
 		Namespace:       "ADMIN",
 		BackfillStreams: []string{"USERS"},
 		CDCStreams:      nil,
@@ -22,7 +22,7 @@ func TestOraclePerformance(t *testing.T) {
 		SupportsCDC:     false,
 	}
 
-	abstract.RunPerformanceTest(t, config)
+	testutils.RunPerformanceTest(t, config)
 }
 
 func connectDatabase(ctx context.Context) (interface{}, error) {
