+++ drivers/postgres/internal/performance_test.go
@@ -5,15 +5,15 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/datazip-inc/olake/drivers/abstract"
 	"github.com/datazip-inc/olake/utils"
+	"github.com/datazip-inc/olake/utils/testutils"
 	"github.com/jmoiron/sqlx"
 	_ "github.com/lib/pq"
 )
 
 func TestPostgresPerformance(t *testing.T) {
-	config := abstract.PerformanceTestConfig{
-		TestConfig:      abstract.GetTestConfig("postgres"),
+	config := testutils.PerformanceTestConfig{
+		TestConfig:      testutils.GetTestConfig("postgres"),
 		Namespace:       "public",
 		BackfillStreams: []string{"test"},
 		CDCStreams:      []string{"test_cdc"},
@@ -24,7 +24,7 @@ func TestPostgresPerformance(t *testing.T) {
 		SupportsCDC:     true,
 	}
 
-	abstract.RunPerformanceTest(t, config)
+	testutils.RunPerformanceTest(t, config)
 }
 
 func connectDatabase(ctx context.Context) (interface{}, error) {
